Fix yellow tint in noise output.
Jakub Beránek
1 month ago (edited)
The yellow tint in the random color generation and the extra ff in memory
is caused by the fact that you are doing bitwise shifts on the char type.
When you do red << 16, it promotes the red to int, so 0xFF becomes
0xFFFFFFFF and that's where the extra FF at the highest byte comes from.
You should probably change the chars to integers, and work with unsigned
integers, that should be safer when doing bitwise shifts and memory
operations in general. The yellow is there because if the green (or blue)
contain a 1 in their MSB, they get filled with ones after promotion to int
because of two's complement and they "leak" into the upper colors
:-)